shrink nrmessage
