+++ src/util/nrmessage.c
@@ -13,7 +13,6 @@
 
 #include <platform.h>
 #include "nrmessage.h"
-#include "nrmessage_struct.h"
 
 /* util includes */
 #include "log.h"
@@ -26,6 +25,13 @@
 #include <assert.h>
 #include <string.h>
 #include <stdlib.h>
+
+typedef struct nrmessage_type {
+  const struct message_type *mtype;
+  char *vars;
+  struct nrmessage_type *next;
+  const char *section;
+} nrmessage_type;
 
 #define NRT_MAXHASH 1021
 static nrmessage_type *nrtypes[NRT_MAXHASH];
